Moved errors.h to base/ ; updated some dependencies.
